# packages/shared-contracts/openapi/cloudhelm.openapi.yaml
openapi: 3.1.0
info:
  title: CloudHelm Platform API
  description: CloudHelm 平台 API。M2 提供真实数据库驱动的项目、任务、需求、设计、审批和事件底座。
  version: 0.2.0
paths:
...
  /api/projects:
    post:
      tags:
      - projects
      summary: 创建项目
      description: 创建 Project，并在同一事务中追加 ProjectCreated 事件。
      operationId: create_project_api_projects_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - projects
      summary: 分页读取项目
      description: 分页读取项目列表。
      operationId: list_projects_api_projects_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 分页大小。
          default: 50
          title: Limit
        description: 分页大小。
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: offset cursor。
          title: Cursor
        description: offset cursor。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_ProjectRead_'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/projects/{project_id}:
    get:
      tags:
      - projects
      summary: 读取项目详情
      description: 读取单个 Project。
      operationId: get_project_api_projects__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/tasks/{task_id}/technical-designs:
    post:
      tags:
      - technical-designs
      summary: 创建技术设计
      description: 保存真实技术设计内容，不触发自动 Architect Agent。
      operationId: create_design_api_tasks__task_id__technical_designs_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicalDesignCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalDesignRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - technical-designs
      summary: 分页读取任务技术设计
      description: 读取某任务下的技术设计列表。
      operationId: list_designs_api_tasks__task_id__technical_designs_get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 分页大小。
          default: 50
          title: Limit
        description: 分页大小。
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: offset cursor。
          title: Cursor
        description: offset cursor。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_TechnicalDesignRead_'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/technical-designs/{design_id}:
    get:
      tags:
      - technical-designs
      summary: 读取技术设计
      description: 读取单个技术设计。
      operationId: get_design_api_technical_designs__design_id__get
      parameters:
      - name: design_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Design Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalDesignRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/technical-designs/{design_id}/approve:
    post:
      tags:
      - technical-designs
      summary: 通过技术设计
      description: 通过技术设计并写入 TechnicalDesignApproved 事件。
      operationId: approve_design_api_technical_designs__design_id__approve_post
      parameters:
      - name: design_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Design Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DecisionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalDesignRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/technical-designs/{design_id}/request-changes:
    post:
      tags:
      - technical-designs
      summary: 要求修改技术设计
      description: 要求修改技术设计并写入事件。
      operationId: request_design_changes_api_technical_designs__design_id__request_changes_post
      parameters:
      - name: design_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Design Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DecisionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalDesignRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    DecisionRequest:
      properties:
        actor_id:
          type: string
          minLength: 1
          title: Actor Id
          description: 操作人或组件标识。
          default: user
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
          description: 操作原因。
      type: object
      title: DecisionRequest
      description: 审批、通过或要求修改类动作的请求体。
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
          description: 稳定错误码。
        message:
          type: string
          title: Message
          description: 面向调用方的错误描述。
        detail:
          anyOf:
          - {}
          - type: 'null'
          title: Detail
          description: 调试细节或字段错误。
        trace_id:
          type: string
          title: Trace Id
          description: 链路追踪 ID。
      type: object
      required:
      - code
      - message
      - trace_id
      title: ErrorResponse
      description: 统一错误响应结构。
...
    PageInfo:
      properties:
        limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Limit
          description: 本次最多返回条数。
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: 下一页游标；为空表示无更多数据。
      type: object
      required:
      - limit
      title: PageInfo
      description: '分页元数据。


        M2 使用 offset cursor 的最小实现，后续如需高并发滚动查询可升级为

        基于 `created_at,id` 的稳定游标。'
...
    PageResponse_ProjectRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectRead'
          type: array
          title: Items
          description: 当前页数据。
        page:
          $ref: '#/components/schemas/PageInfo'
          description: 分页信息。
      type: object
      required:
      - items
      - page
      title: PageResponse[ProjectRead]
...
    PageResponse_TechnicalDesignRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TechnicalDesignRead'
          type: array
          title: Items
          description: 当前页数据。
        page:
          $ref: '#/components/schemas/PageInfo'
          description: 分页信息。
      type: object
      required:
      - items
      - page
      title: PageResponse[TechnicalDesignRead]
...
    ProjectCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
          description: 项目显示名称。
        repo_url:
          type: string
          minLength: 1
          title: Repo Url
          description: 仓库地址或本地仓库路径。
        default_branch:
          type: string
          minLength: 1
          title: Default Branch
          description: 默认分支。
          default: main
        provider:
          type: string
          minLength: 1
          title: Provider
          description: 仓库提供方。
          default: git
      type: object
      required:
      - name
      - repo_url
      title: ProjectCreate
      description: 创建 Project 的请求体。
    ProjectRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        repo_url:
          type: string
          title: Repo Url
        default_branch:
          type: string
          title: Default Branch
        provider:
          type: string
          title: Provider
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - name
      - repo_url
      - default_branch
      - provider
      - created_at
      - updated_at
      title: ProjectRead
      description: Project 响应结构。
...
    ReviewStatus:
      type: string
      enum:
      - draft
      - approved
      - changes_requested
      title: ReviewStatus
      description: 需求和技术设计的人工评审状态。
    RiskLevel:
      type: string
      enum:
      - L0
      - L1
      - L2
      - L3
      - L4
      title: RiskLevel
      description: CloudHelm 工具和任务风险等级。
...
    TechnicalDesignCreate:
      properties:
        requirement_spec_id:
          type: string
          format: uuid
          title: Requirement Spec Id
          description: 关联需求规格 ID。
        design_type:
          type: string
          minLength: 1
          title: Design Type
          description: 设计类型。
          default: mvp
        content_markdown:
          type: string
          minLength: 1
          title: Content Markdown
          description: 技术设计正文。
        openapi_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Openapi Json
          description: OpenAPI 草案 JSON。
        db_schema_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Db Schema Json
          description: 数据库 schema 草案 JSON。
        mermaid_diagram:
          anyOf:
          - type: string
          - type: 'null'
          title: Mermaid Diagram
          description: Mermaid 图。
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
          description: 设计涉及风险等级。
          default: L0
        created_by_agent_run_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Created By Agent Run Id
          description: 创建该设计的 AgentRun。
      type: object
      required:
      - requirement_spec_id
      - content_markdown
      title: TechnicalDesignCreate
      description: 创建技术设计的请求体。
    TechnicalDesignRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        task_id:
          type: string
          format: uuid
          title: Task Id
        requirement_spec_id:
          type: string
          format: uuid
          title: Requirement Spec Id
        design_type:
          type: string
          title: Design Type
        content_markdown:
          type: string
          title: Content Markdown
        openapi_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Openapi Json
        db_schema_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Db Schema Json
        mermaid_diagram:
          anyOf:
          - type: string
          - type: 'null'
          title: Mermaid Diagram
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        status:
          $ref: '#/components/schemas/ReviewStatus'
        created_by_agent_run_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Created By Agent Run Id
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - task_id
      - requirement_spec_id
      - design_type
      - content_markdown
      - openapi_json
      - db_schema_json
      - mermaid_diagram
      - risk_level
      - status
      - created_by_agent_run_id
      - version
      - created_at
      - updated_at
      title: TechnicalDesignRead
      description: 技术设计响应结构。
...
servers:
- url: http://127.0.0.1:18080
  description: 本地开发平台 API